Rewrite CramFS handler so that it inherit the new Handler base class.
